requirements.txt
python-telegram-bot==20.8
web3==6.10.0
